backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-Mail
Flask-CORS
Flask-PyMongo
python-decouple
PyMuPDF
docx2txt
scikit-learn
sentence-transformers
pymongo
dnspython
requests
itsdangerous
werkzeug
google-generativeai